server/requirements.txt
mysql-connector-python
pydantic
uvicorn[standard]
python-dotenv
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
